+++ source/timers.hpp
@@ -18,8 +18,8 @@ inline int64_t time_diff_ms(uint64_t f, uint64_t s) {
 	return (s - f) / TB_TIMER_CLOCK;
 }
 
-inline float time_diff_s(uint64_t f, uint64_t s) {
-	return float(double(s - f) / double(1000UL * TB_TIMER_CLOCK));
+inline double time_diff_s(uint64_t f, uint64_t s) {
+	return double(s - f) / double(1000UL * TB_TIMER_CLOCK);
 }
 
 #endif
